add integrator/comb reset
Reset (zero) the integrator and comb section registers
when the data buffer is reset between user switches in
the case of a single waterfall shared amongst multiple
users (Spartan 6). Done in an attempt to fix the
waterfall contamination problem for this case. It didn't help
so may remove these changes later, although when
unused the logic should get optimized away.