bug fixes, code cleanup, add debug code
Now that public is starting to use prototype all sorts of
bugs are starting to appear. Some easy to fix. Some
requiring debug code added to print more info when the
failing conditions occur.
Tried to add a timer showing user connect time. But this
started causing what looked like memory smashes which
are extremely hard to debug. Save this for later.