Don't disable the /status HTTP query
While the intent seems to be that it's internally used by sdr.hu, and
therefore seemingly unnecessary when the receiver is not registered
there, it blocks off the only means receiver metadata is provided in a
machine-friendly format. This makes it much more difficult and/or less
user-friendly for native apps to create offline lists of receivers not
published at sdr.hu.

Of note is that this restriction is not present in vanilla OpenWebRX.