Enhance weak signals on the waterfall
This patch stretches out the weak signals on the waterfall along
the color palette, and compresses the strong signals a little bit.
Medium signals get moved up the color scale a little, so the
waterfall color may change for some people.

This makes weak signals more easily visible on the waterfall,
and makes adjustment of "WF min" less critical.

I can now see SSB signals on the waterfall that are almost
inaudible, where before I was able to listen to SSB signals
that were invisible on the waterfall.