release v1.21: fix Firefox silence bug detector
The workaround that detects the Firefox bug where audio randomly goes
silent was broken a while back by an unrelated change.
This release also saves AGC parameters in browser cookies, and the
state of the "more" button. If you shift-click the AGC button the
parameters are restored to their default values.