add zoom levels 12-15 (now 16 levels total)
The zoom “max in” button toggles between a nominal max zoom (=z10
currently) to max in (=z15). This is because both are useful and z15 is
really overkill for most uses. There isn’t zoom to add another zoom
button. The button color is orange for z10 and the max in/out buttons
are green when at their extremes.