timestamp audio IQ samples
Every FPGA buffer of audio IQ samples is timestamped with the ADC clock
“ticks” counter. This can be translated into GPS time since the same
ticks counter is captured by the GPS when a position/time solution is
computed. These are just the FPGA and low-level changes. More work
needs to be done to present the timestamp information along with the IQ
samples to the extension interface.