fix .js file caching
Fix caching of .js files. Problem is that the browser might be caching
a .js when the server is now appending an updated version number to it
due to a server recompilation (where the .js file is otherwise
unchanged).

The solution is to return the later of the .js file and server build
time as the time value.