new clock module file
Corrected clock is now stored per-connection so that it might be
disabled on some channels as a future option. MMA changed from 8 to 16
periods. Filter window is 50ppm on first correction for temperature
offset capture but then reduced to 1ppm for subsequent tracking. Fixed
huge bug in construction of 48-bit “ticks” value returned from SPI
data. This was responsible for what was though to be occasional bad GPS
solutions.