v1.99: show sdr.hu status, add admin console
The sdr.hu tab of the admin page now shows the status returned by
sdr.hu when the Kiwi attempts to register. This will give some insight
when the registration process fails. Previously you had to go look for
a message in the log.

The new console tab allows you to control a Beagle Debian shell
directly from the admin page. This is an alternative to using SSH or
PuTTY to login to the Beagle. It of course requires that the Kiwi
server is successfully running.