v1.100: FAX image shear greatly improved
The FAX image shear/slant/skew/tilt, whatever you want to call it, is
now properly matched to the corrected ADC clock. It will still be very
slightly off and there will be an adjustment for this in the future.

This means it is more important than ever to get, and keep, your Kiwi
on-frequency. Either by setting up the GPS antenna properly and having
it correct the ADC clock or by manually trimming it yourself using the
entry on the new right-click menu (see config tab of admin page for
instructions).

Large FAX image captures can now be taken without causing a crash. The
download afterwards is slow and needs to be improved.