add DNS_lookup(), ip_match()
Store both ipv4 and ipv6 ip addresses when doing a DNS lookup. This is
because of ip address comparisons, e.g. to sdr.hu, for Kiwis on
ipv6-enabled networks.

Also, fix improper use of NI_MAXHOST which is the max length of a
hostname and is huge (1025 bytes). Define our own NET_ADDRSTRLEN which
is 64 to fit the usual INET6_ADDRSTRLEN (46) plus some extras in case
ipv6 scope/zone info is appended.