allow passbands down to 4 Hz wide
At 4 Hz the filter skits aren’t so great as a percentage of passband
width. But there is a S/N improvement over the previous minimum of 50
Hz. So this is a worthwhile change.

Also, the passband narrow/widen will now saturate at the min/max values
instead of stopping short.