debugging for strange "subframe 7" seen
In the past we’ve occasionally seen subframe = 0 which is an illegal
value. But all of a sudden most Kiwis started seeing subframe = 7 on
multiple PRNs. By the time the debugging code was developed the problem
has disappeared. If the problem reoccurs the debug code will determine
where in the transmission sequence the bad subframes are seen, and
perhaps if there is any consistency.