show msg while determining local interface
When server (re)starts there is a short period of time before the ip of
the local interface is determined. If there is an admin connection
during this time the “local connection” auto-login cannot be
determined. If a password had been set then we showed the password
panel. But this proves to be confusing to people. So instead we show a
message saying the address is still being determined and to reload the
page in a few moments.