clean termination of websocket connections kicked out by the server
Before the SND (and WF) process just stopped to send data to the websocket
server. Because of this, mongoose sent a ping packet after a timeout
of 5 sec, expecting an immediate answer from the client and then closed the connection.
But the client had no way of knowing that its connection was being closed by the server.

Now the server sends a close request packet to the client which allows to
terminate the websocket connection in a clean way.