Debian 7 compatibility
There are only 6 Beta test Kiwis left running Debian 7. At some point
we really need to get them upgraded, but this is complicated because
we’d like to be able to do this in an automated way not requiring
manual intervention. This is important to figure out for when the Kiwi
user base in general needs to migrate from Debian 8 to something more
recent.