always send DDNS update to noip.com
Some people think there is a problem with the Kiwi DUC because the
noip.com status page “last update” time never seems to change. What
they don’t realize is that it only updates if the ip address has
actually changed. Not for each time the DUC runs and does a check.

To reduce this confusion the DUC will now forcibly send an update to
noip.com whether the ip address has changed or not.