Update sats.cpp to activate QZS-1R L1 C/A signal
Uncommented the satellite data of PRN 196 so that we can make use of QZS-1R (replacement) that was launched on 2021-10-26. The alert flag will be unset on 2022-01-26 (https://qzss.go.jp/info/information/qzs1r_211222.html and the information is in Japanese only...) and the QZS-1R will be available on 2022-03-03.

SVIDs in QZSS were changed because there are difference between the SVN (space vehicular number) and the SVID (space vehicular identification). For example, the SVID of QZS-4 is 3 and it is denoted J03, say, in  RINEX (page 25 on IS-QZSS-PNT https://qzss.go.jp/en/technical/download/pdf/ps-is-qzss/is-qzss-pnt-004.pdf?t=1641020780449).